--- FileSystem/cluster/data_cluster.c
@@ -1,3 +1,0 @@
-#include <stdlib.h>
-#include <stdio.h>
-#include "data_cluster.h"
