--- FileSystem/dir/dir_entry.c
@@ -1,3 +1,0 @@
-#include <stdlib.h>
-#include <stdio.h>
-#include "dir_entry.h"
