+++ FileSystem/shell/init/init_file_system.h
@@ -1,0 +1,8 @@
+#ifndef FILESYSTEM_INIT_FILE_SYSTEM_H
+#define FILESYSTEM_INIT_FILE_SYSTEM_H
+
+#include <stdint.h>
+
+FILE* initializateFileSystem(uint16_t* fat);
+
+#endif //FILESYSTEM_INIT_FILE_SYSTEM_H
