+++ FileSystem/shell/shell.h
@@ -1,0 +1,6 @@
+#ifndef FILESYSTEM_SHELL_H
+#define FILESYSTEM_SHELL_H
+
+void handleUserInput(const char* userInput);
+
+#endif //FILESYSTEM_SHELL_H
