--- FileSystem/utils.c
@@ -1,7 +1,0 @@
-#include <stdlib.h>
-#include <stdbool.h>
-#include "utils.h"
-
-bool starts_with(const char *pre, const char *str) {
-    return strncmp(pre, str, strlen(pre)) == 0;
-}
