--- FileSystem/utils.h
@@ -1,8 +1,0 @@
-#ifndef FILESYSTEM_UTILS_H
-#define FILESYSTEM_UTILS_H
-
-#include <stdbool.h>
-
-bool starts_with(const char *pre, const char *str);
-
-#endif //FILESYSTEM_UTILS_H
